requirements.txt
-i https://pypi.org/simple
--extra-index-url https://pypi.org/simple
boto3
elementpath
datetime
mysql-connector